组织机构+老师数据库
git-svn-id: http://192.168.1.227/svn/CMS4j/trunk@334 d1d3f527-36cb-4749-94ba-585de1865e2b

+++ src/main/java/cn/edu/sdufe/cms/common/entity/agency/Agency.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component;
 public class Agency extends PersistableEntity {
     private String title;
 
-    private int categoryId;
+    private Long categoryId;
 
     private String imageUrl;
 
@@ -32,11 +32,11 @@ public class Agency extends PersistableEntity {
         this.title = title;
     }
 
-    public int getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(int categoryId) {
+    public void setCategoryId(Long categoryId) {
         this.categoryId = categoryId;
     }
